DSS-320: Feature plot tank layout (#143)
* entity class change

* adding loadable pattern tank details

* Generate tank layout

* test cases updation

* code review comments fixing

* spotless apply

+++ api/cp-dss-api/vessel-info/src/main/java/com/cpdss/vesselinfo/entity/InnerBulkHeadValues.java
@@ -23,7 +23,7 @@ public class InnerBulkHeadValues extends EntityDoc {
   private static final long serialVersionUID = 1L;
 
   @Column(name = "vessel_xid")
-  private Integer vesselId;
+  private Long vesselId;
 
   @Column(name = "frame_number")
   private BigDecimal frameNumber;
